Refactor hash map initialization and resizing functions for improved clarity and safety

+++ hash.h
@@ -14,7 +14,8 @@ void print_hash_32(const char* str) {
 void print_hash_128_x86(const char* str) {
     uint32_t hash[4];
     MurmurHash3_x86_128(str, strlen(str), 0, hash);
-    printf("MurmurHash3_x86_128(\"%s\") = 0x%08x%08x%08x%08x\n", str, hash[0], hash[1], hash[2], hash[3]);
+    printf("MurmurHash3_x86_128(\"%s\") = 0x%08x%08x%08x%08x\n", str, hash[0], hash[1], hash[2],
+hash[3]);
 }
 
 void print_hash_128_x64(const char* str) {
@@ -119,9 +120,9 @@ static inline uint32_t MurmurHash3_x86_32(const void* key, int len, uint32_t see
 
     switch (len & 3) {
         case 3:
-            k1 ^= tail[2] << 16;
+            k1 ^= (uint32_t)tail[2] << 16;
         case 2:
-            k1 ^= tail[1] << 8;
+            k1 ^= (uint32_t)tail[1] << 8;
         case 1:
             k1 ^= tail[0];
             k1 *= c1;
@@ -131,7 +132,7 @@ static inline uint32_t MurmurHash3_x86_32(const void* key, int len, uint32_t see
     }
 
     // Finalization
-    h1 ^= len;
+    h1 ^= (uint32_t)len;
     h1 = fmix32(h1);
 
     return h1;
@@ -202,50 +203,50 @@ static inline void MurmurHash3_x86_128(const void* key, const int len, uint32_t 
 
     switch (len & 15) {
         case 15:
-            k4 ^= tail[14] << 16;
+            k4 ^= (uint32_t)tail[14] << 16;
         case 14:
-            k4 ^= tail[13] << 8;
+            k4 ^= (uint32_t)tail[13] << 8;
         case 13:
-            k4 ^= tail[12] << 0;
+            k4 ^= (uint32_t)tail[12] << 0;
             k4 *= c4;
             k4 = ROTL32(k4, 18);
             k4 *= c1;
             h4 ^= k4;
 
         case 12:
-            k3 ^= tail[11] << 24;
+            k3 ^= (uint32_t)tail[11] << 24;
         case 11:
-            k3 ^= tail[10] << 16;
+            k3 ^= (uint32_t)tail[10] << 16;
         case 10:
-            k3 ^= tail[9] << 8;
+            k3 ^= (uint32_t)tail[9] << 8;
         case 9:
-            k3 ^= tail[8] << 0;
+            k3 ^= (uint32_t)tail[8] << 0;
             k3 *= c3;
             k3 = ROTL32(k3, 17);
             k3 *= c4;
             h3 ^= k3;
 
         case 8:
-            k2 ^= tail[7] << 24;
+            k2 ^= (uint32_t)tail[7] << 24;
         case 7:
-            k2 ^= tail[6] << 16;
+            k2 ^= (uint32_t)tail[6] << 16;
         case 6:
-            k2 ^= tail[5] << 8;
+            k2 ^= (uint32_t)tail[5] << 8;
         case 5:
-            k2 ^= tail[4] << 0;
+            k2 ^= (uint32_t)tail[4] << 0;
             k2 *= c2;
             k2 = ROTL32(k2, 16);
             k2 *= c3;
             h2 ^= k2;
 
         case 4:
-            k1 ^= tail[3] << 24;
+            k1 ^= (uint32_t)tail[3] << 24;
         case 3:
-            k1 ^= tail[2] << 16;
+            k1 ^= (uint32_t)tail[2] << 16;
         case 2:
-            k1 ^= tail[1] << 8;
+            k1 ^= (uint32_t)tail[1] << 8;
         case 1:
-            k1 ^= tail[0] << 0;
+            k1 ^= (uint32_t)tail[0] << 0;
             k1 *= c1;
             k1 = ROTL32(k1, 15);
             k1 *= c2;
@@ -253,10 +254,10 @@ static inline void MurmurHash3_x86_128(const void* key, const int len, uint32_t 
     }
 
     // Finalization
-    h1 ^= len;
-    h2 ^= len;
-    h3 ^= len;
-    h4 ^= len;
+    h1 ^= (uint32_t)len;
+    h2 ^= (uint32_t)len;
+    h3 ^= (uint32_t)len;
+    h4 ^= (uint32_t)len;
 
     h1 += h2;
     h1 += h3;
@@ -286,7 +287,8 @@ static inline void MurmurHash3_x86_128(const void* key, const int len, uint32_t 
 //-----------------------------------------------------------------------------
 // MurmurHash3_x64_128 - 128-bit hash for x64 platforms (returns 2 x 64-bit values)
 //-----------------------------------------------------------------------------
-static inline void MurmurHash3_x64_128(const void* key, const int len, const uint32_t seed, void* out) {
+static inline void MurmurHash3_x64_128(const void* key, const int len, const uint32_t seed,
+                                       void* out) {
     const uint8_t* data = (const uint8_t*)key;
     const int nblocks   = len / 16;
 
@@ -367,8 +369,8 @@ static inline void MurmurHash3_x64_128(const void* key, const int len, const uin
     }
 
     // Finalization
-    h1 ^= len;
-    h2 ^= len;
+    h1 ^= (uint32_t)len;
+    h2 ^= (uint32_t)len;
 
     h1 += h2;
     h2 += h1;
